# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Openpage API",
        "version": "1.0.0"
    },
    "basePath": "/openpage-backend/public",
    "paths": {
...
        "/subject": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "summary": "Returns list of subjects",
                "description": "Returns subject data",
                "operationId": "subject_list",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "search_key",
                        "in": "query",
                        "description": "Search parameter or key word to search",
                        "type": "string"
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "this is offset or skip the records",
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of records to be retrieved ",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subject"
                ],
                "summary": "Update Subject details",
                "description": "",
                "operationId": "create_subject",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Subject json input",
                        "required": true,
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subject"
                ],
                "summary": "Create a Subject",
                "description": "",
                "operationId": "create_subject",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "subject json input",
                        "required": true,
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subject"
                ],
                "summary": "delete subject data",
                "description": "Delete subject from system",
                "operationId": "delete_subject",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "query",
                        "description": "ID of the subject that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid data supplied"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/subject/{_id}": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "summary": "Returns subject data",
                "description": "Returns subject data",
                "operationId": "subject_list",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "description": "ID of the subject that needs to be displayed",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid subject id"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    },
...
}